Rework the generator
This is how it should've been done in the first place, if I went slower.
Claude/composer wasn't smart enough to design things the easy way.

Basically, just like with random_constr.py, while we try to generate a
construction, actually simulate running it and simply discard any
attempted commands that generate errors, so it is guaranteed to be
semantically valid.

This makes things radically simpler, like we don't need a whole massive
function just to handle return values, because return values are
actually returned to us. Likewise, we don't need to simulate commands or
construct command representations, we just construct/apply Command
objects and use its repr().

Accordingly, modify the rest of the program to operate with Element and
Command semantics directly.

Also, separate the logic to sample the command out from the logic to
execute the command. This will help generate specific constructions
later.

This probably had to be done, even considering the short time horizon of
the project, because the bug it induced was really critical-- we weren't
handling retvals correctly, so lots of more interesting constructions,
including most of the ones with polygons, were just not workable.