Add statistics to translated problem JSON
This is slightly messy.
In a way the most elegant thing to do is to merge the whole pipeline
into one while the problem is still being constructed, in one big
script.
That way we would have all the semantic data about a problem all in one
place, and we could construct a JSON out of that and linking to the
command language version of the problem would be instant.

Because I don't want to do all that refactoring work for that little
feature, currently the sorts of stats we can collect are limited, and
also some hacks are needed to gather certain stats.