vLLM optimizations to the grader
found that openthinker was not using kv cache

attempting to optimize more will probably not work.
because openthinker is a reasoning model, its sequences are long, and
runtime is dominated by attention calculation, not prefill, and
attention calculation is already maximally optimized by vllm, using
flashattention 3 and selecting the best kernel for the architecture
automatically.

see dev notes.